+++ docs/demo/bootstrap/bundle.js
@@ -342,10 +342,11 @@ function textbox(opts, ctx) {
       if (transformer) {
         value = transformer.parse(value);
       }
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -414,10 +415,11 @@ function checkbox(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -502,10 +504,11 @@ function select(opts, ctx) {
       if (value === nullOption.value) {
         value = null;
       }
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -572,10 +575,11 @@ function radio(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -671,10 +675,11 @@ function struct(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -799,10 +804,11 @@ function list(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1458,7 +1464,7 @@ function getHiddenTextbox(locals) {
     },
     events: {
       change: function (evt) {
-        locals.onChange(evt.target.value)
+        locals.onChange(evt.target.value);
       }
     }
   };
@@ -1479,7 +1485,7 @@ function textbox(locals) {
     'aria-describedby': locals.help ? locals.id + '-tip' : null,
     id: locals.label ? locals.id : null,
     onChange: function (evt) {
-      locals.onChange(evt.target.value)
+      locals.onChange(evt.target.value);
     },
     placeholder: locals.placeholder,
     name: locals.name,
@@ -1654,7 +1660,7 @@ function radio(locals) {
       label: option.text,
       name: locals.name,
       onChange: function (evt) {
-        locals.onChange(evt.target.value)
+        locals.onChange(evt.target.value);
       },
       value: option.value
     });