+++ docs/demo/ionic/bundle.js
@@ -207,10 +207,11 @@ function textbox(opts, ctx) {
       if (transformer) {
         value = transformer.parse(value);
       }
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -279,10 +280,11 @@ function checkbox(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -367,10 +369,11 @@ function select(opts, ctx) {
       if (value === nullOption.value) {
         value = null;
       }
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -437,10 +440,11 @@ function radio(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -536,10 +540,11 @@ function struct(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -664,10 +669,11 @@ function list(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1323,7 +1329,7 @@ function getHiddenTextbox(locals) {
     },
     events: {
       change: function (evt) {
-        locals.onChange(evt.target.value)
+        locals.onChange(evt.target.value);
       }
     }
   };
@@ -1344,7 +1350,7 @@ function textbox(locals) {
     'aria-describedby': locals.help ? locals.id + '-tip' : null,
     id: locals.label ? locals.id : null,
     onChange: function (evt) {
-      locals.onChange(evt.target.value)
+      locals.onChange(evt.target.value);
     },
     placeholder: locals.placeholder,
     name: locals.name,
@@ -1519,7 +1525,7 @@ function radio(locals) {
       label: option.text,
       name: locals.name,
       onChange: function (evt) {
-        locals.onChange(evt.target.value)
+        locals.onChange(evt.target.value);
       },
       value: option.value
     });
@@ -1703,7 +1709,7 @@ function textbox(locals) {
         },
         events: {
           change: function (evt) {
-            locals.onChange(evt.target.value)
+            locals.onChange(evt.target.value);
           }
         }
       }