+++ docs/guide/bundle.js
@@ -868,10 +868,11 @@ function textbox(opts, ctx) {
       if (transformer) {
         value = transformer.parse(value);
       }
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -940,10 +941,11 @@ function checkbox(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1028,10 +1030,11 @@ function select(opts, ctx) {
       if (value === nullOption.value) {
         value = null;
       }
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1098,10 +1101,11 @@ function radio(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1197,10 +1201,11 @@ function struct(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1325,10 +1330,11 @@ function list(opts, ctx) {
     },
 
     onChange: function (value) {
+      this.state.value = value;
+      this.forceUpdate();
       if (this.props.onChange) {
         this.props.onChange(value);
       }
-      this.setState({value: value});
     },
 
     getValue: function () {
@@ -1984,7 +1990,7 @@ function getHiddenTextbox(locals) {
     },
     events: {
       change: function (evt) {
-        locals.onChange(evt.target.value)
+        locals.onChange(evt.target.value);
       }
     }
   };
@@ -2005,7 +2011,7 @@ function textbox(locals) {
     'aria-describedby': locals.help ? locals.id + '-tip' : null,
     id: locals.label ? locals.id : null,
     onChange: function (evt) {
-      locals.onChange(evt.target.value)
+      locals.onChange(evt.target.value);
     },
     placeholder: locals.placeholder,
     name: locals.name,
@@ -2180,7 +2186,7 @@ function radio(locals) {
       label: option.text,
       name: locals.name,
       onChange: function (evt) {
-        locals.onChange(evt.target.value)
+        locals.onChange(evt.target.value);
       },
       value: option.value
     });